BalatroDocker/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-uinput
requests
fastmcp
pyautogui
Pillow